Move theme colour setting to layout
